feat(motor-control): implement RPM-based motor control system
- Replace PWM motor control with RPM-based system for better precision
- Add PID synchronization between master and slave controllers
- Introduce new motor settings menu for RPM and PID configuration
- Implement serial communication protocol for RPM commands
- Add safety checks and error handling for motor operations
- Update documentation with new motor control guidelines

+++ agv_sami/pid.ino
@@ -1,4 +1,4 @@
-double computePID(int index, double setpoint, double input, double Kp, double Ki, double Kd, int Minintegral, double Maxintegral) {
+double computePID(int index, double setpoint, double input, double Kp, double Ki, double Kd, double Minintegral, double Maxintegral) {
   // Validate input parameters
   if (index < 0 || index >= numOutputs) {
     logError(ERROR_PID_CALCULATION, "PID index tidak valid");
